+++ mat-file-io/src/main/java/us/hebi/matlab/common/memory/ByteConverters.java
@@ -41,28 +41,13 @@ public class ByteConverters {
         return heapConverter;
     }
 
-    /**
-     * Native access that uses sun.misc.unsafe to read and write to arbitrary memory
-     * locations such as the backing memory of native buffers and mapped files. This
-     * may not perform any bounds checks and may segfault. Use with caution!
-     *
-     * @param atomic            makes reads/writes act like volatile variables
-     * @param alwaysNativeOrder ignores the desired order and always assumes native order
-     * @param baseOffset        base offset that gets added to any specified offset. Default=0
-     */
-    public static NativeAccess getNativeAccess(boolean atomic, boolean alwaysNativeOrder, long baseOffset) {
-        UnsafeAccess.requireUnsafe();
-        NativeAccess base = atomic ? new NativeAtomicAccess(baseOffset) : new NativeRegularAccess(baseOffset);
-        return !alwaysNativeOrder ? new NativeOrderConverter(base) : base;
-    }
-
     static {
         heapConverter = new HeapByteConverter();
 
         if (UnsafeAccess.isAvailable()) {
             final UnsafeByteConverter unsafeConverter = new UnsafeByteConverter();
             rawUnsafeConverter = unsafeConverter;
-            unsafeConverterWithBoundsCheck = new CheckArrayBounds(unsafeConverter);
+            unsafeConverterWithBoundsCheck = new ArrayBoundsCheck(unsafeConverter);
         } else {
             rawUnsafeConverter = null;
             unsafeConverterWithBoundsCheck = null;
